Import chromium-browser_50.0.2661.102.orig.tar.xz from Ubuntu Yakkety
All .gitignore files dropped, as this would have prevented much of the
tarball being imported.
 # find . -name .gitignore -delete

Removed files over 50MB in size to make GitHub happy.
 # find . -type f -size +49M -delete

+++ chrome/android/java/src/org/chromium/chrome/browser/ntp/RecentTabsManager.java
@@ -5,7 +5,9 @@
 package org.chromium.chrome.browser.ntp;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.graphics.Bitmap;
+import android.preference.PreferenceManager;
 
 import org.chromium.base.ObserverList;
 import org.chromium.base.ThreadUtils;
@@ -22,9 +24,10 @@ import org.chromium.chrome.browser.ntp.RecentTabsPromoView.SyncPromoModel;
 import org.chromium.chrome.browser.ntp.RecentlyClosedBridge.RecentlyClosedCallback;
 import org.chromium.chrome.browser.ntp.RecentlyClosedBridge.RecentlyClosedTab;
 import org.chromium.chrome.browser.profiles.Profile;
+import org.chromium.chrome.browser.signin.SigninAccessPoint;
 import org.chromium.chrome.browser.signin.SigninManager;
 import org.chromium.chrome.browser.signin.SigninManager.SignInStateObserver;
-import org.chromium.chrome.browser.sync.SyncController;
+import org.chromium.chrome.browser.sync.ProfileSyncService;
 import org.chromium.chrome.browser.tab.Tab;
 import org.chromium.content_public.browser.LoadUrlParams;
 import org.chromium.sync.AndroidSyncSettings;
@@ -51,6 +54,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
     }
 
     private static final int RECENTLY_CLOSED_MAX_TAB_COUNT = 5;
+    private static final String PREF_SIGNIN_PROMO_DECLINED =
+            "recent_tabs_signin_promo_declined";
 
     private final Profile mProfile;
     private final Tab mTab;
@@ -218,6 +223,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      */
     public void openForeignSessionTab(ForeignSession session, ForeignSessionTab tab,
             int windowDisposition) {
+        if (mIsDestroyed) return;
         NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_FOREIGN_SESSION);
         mForeignSessionHelper.openForeignSessionTab(mTab, session, tab, windowDisposition);
     }
@@ -230,6 +236,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      *         be restored into the current tab or a new tab.
      */
     public void openRecentlyClosedTab(RecentlyClosedTab tab, int windowDisposition) {
+        if (mIsDestroyed) return;
         NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_RECENTLY_CLOSED_ENTRY);
         mRecentlyClosedBridge.openRecentlyClosedTab(mTab, tab, windowDisposition);
     }
@@ -238,6 +245,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Opens the history page.
      */
     public void openHistoryPage() {
+        if (mIsDestroyed) return;
         mTab.loadUrl(new LoadUrlParams(UrlConstants.HISTORY_URL));
         StartupMetrics.getInstance().recordOpenedHistory();
     }
@@ -282,6 +290,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the currently open tabs list is collapsed.
      */
     public void setCurrentlyOpenTabsCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setCurrentlyOpenTabsCollapsed(isCollapsed);
     }
 
@@ -325,6 +334,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the session is collapsed or expanded.
      */
     public void setForeignSessionCollapsed(ForeignSession session, boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setForeignSessionCollapsed(session, isCollapsed);
     }
 
@@ -345,6 +355,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the recently closed tabs list is collapsed.
      */
     public void setRecentlyClosedTabsCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setRecentlyClosedTabsCollapsed(isCollapsed);
     }
 
@@ -365,6 +376,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param session Session to be deleted.
      */
     public void deleteForeignSession(ForeignSession session) {
+        if (mIsDestroyed) return;
         mForeignSessionHelper.deleteForeignSession(session);
     }
 
@@ -372,6 +384,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Clears the list of recently closed tabs.
      */
     public void clearRecentlyClosedTabs() {
+        if (mIsDestroyed) return;
         mRecentlyClosedBridge.clearRecentlyClosedTabs();
     }
 
@@ -385,7 +398,22 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
             return false;
         }
 
+        if (PreferenceManager.getDefaultSharedPreferences(mContext).getBoolean(
+                PREF_SIGNIN_PROMO_DECLINED, false)) {
+            return false;
+        }
+
         return !AndroidSyncSettings.isSyncEnabled(mContext) || mForeignSessions.isEmpty();
+    }
+
+    /**
+     * Save that user tapped "No" button on the signin promo.
+     */
+    public void setSigninPromoDeclined() {
+        SharedPreferences.Editor sharedPreferencesEditor =
+                PreferenceManager.getDefaultSharedPreferences(mContext).edit();
+        sharedPreferencesEditor.putBoolean(PREF_SIGNIN_PROMO_DECLINED, true);
+        sharedPreferencesEditor.apply();
     }
 
     /**
@@ -394,6 +422,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the sync promo is collapsed.
      */
     public void setSyncPromoCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setSyncPromoCollapsed(isCollapsed);
     }
 
@@ -452,7 +481,10 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
 
     @Override
     public void enableSync() {
-        SyncController.get(mContext).start();
+        ProfileSyncService syncService = ProfileSyncService.get();
+        if (syncService != null) {
+            syncService.requestStart();
+        }
     }
 
     @Override
@@ -471,5 +503,10 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
             mProfileDataCache = new ProfileDataCache(mContext, Profile.getLastUsedProfile());
         }
         return mProfileDataCache;
+    }
+
+    @Override
+    public int getAccessPoint() {
+        return SigninAccessPoint.RECENT_TABS;
     }
 }
